feat(security): Add helmet and mongo-sanitize middleware for enhanced security refactor(services): Move getFullQuote to quote.service and update imports feat(quote-stream): Implement message processing for updated quotes in WebSocket
