+++ libs/server/features/portfolios/src/commands/positions/patch-position.handler.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import positionsSchema from '../../schemas/positions.schema';
 import { StatusCodes } from 'http-status-codes';
-import { getFullQuote } from '../../services/fmp/full-quote.service';
+import { getFullQuote } from '../../services/fmp/quote.service';
 
 export const patchPositionHandler = async (req: Request, res: Response) => {
   const user = process.env['NX_PUBLIC_DEV_USER'] ?? '';
