+++ package.json
@@ -22,10 +22,13 @@
     "date-fns-tz": "^3.2.0",
     "dotenv": "^16.4.5",
     "express": "^4.18.1",
+    "express-mongo-sanitize": "^2.2.0",
     "express-oauth2-jwt-bearer": "^1.6.0",
+    "express-rate-limit": "^7.4.1",
     "express-validator": "^7.2.0",
     "express-ws": "^5.0.2",
     "formik": "^2.4.6",
+    "helmet": "^8.0.0",
     "http-status-codes": "^2.3.0",
     "lodash-es": "^4.17.21",
     "mongoose": "^8.6.4",
